perf(blazwitcher): :zap: optimize code style
