+++ sidepanel/index.tsx
@@ -1,33 +1,12 @@
 import './sidepanel.css'
 
 import { Layout } from '@douyinfe/semi-ui'
-import { useAtom } from 'jotai'
 import { useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
 
-import {
-	DEFAULT_BOOKMARK_DISPLAY_COUNT,
-	DEFAULT_HISTORY_DISPLAY_COUNT,
-	DEFAULT_STRICTNESS_COEFFICIENT,
-	MAIN_CONTENT_CLASS,
-	MAIN_WINDOW,
-} from '~shared/constants'
-import type { ItemType, ListItemType } from '~shared/types'
-import {
-	isBookmarkItem,
-	isDarkMode,
-	isHistoryItem,
-	isTabItem,
-	setDarkTheme,
-	splitCompositeHitRanges,
-} from '~shared/utils'
+import { MAIN_CONTENT_CLASS } from '~shared/constants'
+import { orderList, searchWithList, setDarkTheme } from '~shared/utils'
 
-import {
-	type Matrix,
-	isStrictnessSatisfied,
-	mergeSpacesWithRanges,
-	searchSentenceByBoundaryMapping,
-} from 'text-search-engine'
 import Footer from './footer'
 import useOriginalList from './hooks/useOriginalList'
 import List from './list'
@@ -47,62 +26,6 @@ const ContentWrapper = styled(Content)`
   padding: 0;
 `
 
-const orderList = (list: ListItemType[]) => {
-	const tabs: ListItemType<ItemType.Tab>[] = []
-	const bookmarks: ListItemType<ItemType.Bookmark>[] = []
-	const histories: ListItemType<ItemType.History>[] = []
-	const set = new Set()
-	for (const item of list) {
-		const { title, url } = item.data
-		const hasSameItemInTabs = () => {
-			if (set.has(title) || (url && set.has(url))) {
-				return true
-			}
-			set.add(title)
-			set.add(url)
-			return false
-		}
-
-		if (isTabItem(item)) {
-			tabs.push(item)
-			set.add(title)
-			set.add(url)
-		} else if (isBookmarkItem(item) && !hasSameItemInTabs()) {
-			bookmarks.push(item)
-		} else if (isHistoryItem(item) && !hasSameItemInTabs()) {
-			histories.push(item)
-		}
-	}
-	const compareForLastAccess = (a: ListItemType<ItemType.Tab>, b: ListItemType<ItemType.Tab>) =>
-		a.data.lastAccessed ? b.data.lastAccessed - a.data.lastAccessed : -1
-
-	const compareForLastVisitTime = (a: ListItemType<ItemType.History>, b: ListItemType<ItemType.History>) =>
-		a.data.lastVisitTime ? b.data.lastVisitTime - a.data.lastVisitTime : -1
-
-	const compareForHitRangeLength = (a: ListItemType, b: ListItemType) => {
-		if (a.data.compositeHitRanges && b.data.compositeHitRanges) {
-			return a.data.compositeHitRanges.length - b.data.compositeHitRanges.length
-		}
-		return 0
-	}
-
-	const compareForActiveStatus = (a: ListItemType<ItemType.Tab>, _b: ListItemType<ItemType.Tab>) =>
-		a.data.active ? -1 : 1
-
-	return [
-		...tabs
-			.filter((item) => !item.data.url.includes(chrome.runtime.id))
-			.toSorted(compareForLastAccess)
-			.toSorted(compareForHitRangeLength)
-			.toSorted(compareForActiveStatus),
-		...histories
-			.toSorted(compareForLastVisitTime)
-			.toSorted(compareForHitRangeLength)
-			.slice(0, DEFAULT_HISTORY_DISPLAY_COUNT),
-		...bookmarks.toSorted(compareForHitRangeLength).slice(0, DEFAULT_BOOKMARK_DISPLAY_COUNT),
-	].toSorted(compareForHitRangeLength)
-}
-
 export default function SidePanel() {
 	const originalList = useOriginalList()
 	const [searchValue, setSearchValue] = useState('')
@@ -114,24 +37,7 @@ export default function SidePanel() {
 	const RenderList = useMemo(() => {
 		let filteredList = originalList
 		if (searchValue.trim() !== '') {
-			filteredList = originalList.reduce<ListItemType[]>((acc, item) => {
-				let hitRanges: Matrix | undefined
-				hitRanges = searchSentenceByBoundaryMapping(item.data.compositeBoundaryMapping, searchValue)
-				if (hitRanges) {
-					const mergedHitRanges = mergeSpacesWithRanges(item.data.compositeSource, hitRanges)
-					if (isStrictnessSatisfied(DEFAULT_STRICTNESS_COEFFICIENT, searchValue, mergedHitRanges)) {
-						const [titleHitRanges, hostHitRanges] = splitCompositeHitRanges(mergedHitRanges, [
-							item.data.title.length,
-							item.data.host.length,
-						])
-						acc.push({
-							...item,
-							data: { ...item.data, compositeHitRanges: mergedHitRanges, titleHitRanges, hostHitRanges },
-						})
-					}
-				}
-				return acc
-			}, [])
+			filteredList = searchWithList(originalList, searchValue)
 		}
 		// use plugin 后，这里的 prop 时需要被替换掉
 		return <List list={orderList(filteredList)} RenderItem={RenderItem}></List>
